Move side panel into main page dir
