+++ web/src/app/pages/main-page-container/main-page/side-panel/side-panel.component.css
