+++ web/src/app/pages/main-page-container/main-page/side-panel/side-panel.module.ts
@@ -16,7 +16,7 @@
 
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { SidePanelComponent } from 'app/components/side-panel/side-panel.component';
+import { SidePanelComponent } from 'app/pages/main-page-container/main-page/side-panel/side-panel.component';
 import { LocationOfInterestPanelModule } from 'app/components/loi-panel/loi-panel.module';
 import { JobListModule } from 'app/components/job-list/job-list.module';
 import { SubmissionFormModule } from 'app/components/submission-form/submission-form.module';
